creates 3 modules for class environment modules : set_phase, high_five, performance(NOT WORKING FOR TEACHER APPRENTICE) removes common methods shared between teacher and teacherapprentive.
