+++ set_phase.rb
@@ -1,0 +1,13 @@
+module SetPhase
+  def set_phase(num)
+    response = ""
+    if num == @phase
+      response += "I'm doing phase #{@phase} again because "
+      response += "I put my learning first. I'm gonna rock it!"
+    else
+      response = "Oooh, phase #{num}. I hope I'm ready!"
+    end
+    @phase = num
+    response
+  end
+end
