+++ student.rb
@@ -1,3 +1,5 @@
+require_relative 'high_five'
+require_relative 'set_phase'
 class Student
   attr_reader :age, :phase
   attr_accessor :name
@@ -8,21 +10,8 @@ class Student
     @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    response = ""
-    if num == @phase
-      response += "I'm doing phase #{@phase} again because "
-      response += "I put my learning first. I'm gonna rock it!"
-    else
-      response = "Oooh, phase #{num}. I hope I'm ready!"
-    end
-    @phase = num
-    response
-  end
+  include HighFive
+  include SetPhase
 
   def learn_stuff
     response = ""
